+++ R/utils-predicates.R
@@ -1,9 +1,9 @@
 
 is_pure_BackupQueueIndex <- function(
   file,
-  backup_dir = dirname(file)
+  dir = dirname(file)
 ){
-  identical(BackupQueueDateTime$new(file, backup_dir = backup_dir)$n_backups, 0L)
+  identical(BackupQueueDateTime$new(file, dir = dir)$n_backups, 0L)
 }
 
 
@@ -11,9 +11,9 @@ is_pure_BackupQueueIndex <- function(
 
 is_pure_BackupQueueDateTime <- function(
   file,
-  backup_dir = dirname(file)
+  dir = dirname(file)
 ){
-  bi <- BackupQueueIndex$new(file, backup_dir = backup_dir)
+  bi <- BackupQueueIndex$new(file, dir = dir)
   identical(bi$n_backups, 0L) || min(bi$backups$index) > 1L
 }
 
@@ -22,9 +22,9 @@ is_pure_BackupQueueDateTime <- function(
 
 is_pure_BackupQueue <- function(
   file,
-  backup_dir = dirname(file)
+  dir = dirname(file)
 ){
-  bi <- BackupQueueIndex$new(file, backup_dir = backup_dir)
+  bi <- BackupQueueIndex$new(file, dir = dir)
 
   if (bi$n_backups < 1){
     TRUE
@@ -32,7 +32,7 @@ is_pure_BackupQueue <- function(
     # check if min index is 1 to filter out BackupQueueIndex that are truely
     # BackupQueueDate but only have integer like timestamps
     identical(
-      try(BackupQueueDateTime$new(file, backup_dir = backup_dir)$n_backups, silent = TRUE),
+      try(BackupQueueDateTime$new(file, dir = dir)$n_backups, silent = TRUE),
       0L
     )
   } else {
@@ -45,10 +45,10 @@ is_pure_BackupQueue <- function(
 
 assert_pure_BackupQueue <- function(
   file,
-  backup_dir = dirname(file),
+  dir = dirname(file),
   warn_only = FALSE
 ){
-  if (is_pure_BackupQueue(file, backup_dir = backup_dir))
+  if (is_pure_BackupQueue(file, dir = dir))
     return(TRUE)
 
   msg <- paste0(